requirements.txt
openai
pyupbit
pandas
pandas_ta
schedule
python-dotenv